Refactor GetServiceNameFromHost
